refactor(machine): extract the settings state machine into useMachineSettings
SettingsTab had grown to 396 lines carrying 8 pieces of state and 4 refs, and
none of that machinery is about rendering -- it is the load/save/resync state
machine, where every subtlety lives (concurrent saves, per-key revert, deferred
resync, dirty-draft preservation, stale-machine results). Tangled with the form
JSX it reads as incidental complexity; on its own it reads as what it is.

SettingsTab is now 240 lines: the form, plus the delete action. The hook is 216
and documents its own invariants. It lives in apps/web/src/hooks/ beside its
useMachineProjects / useMachineBranches siblings -- deliberately NOT on SWR like
they are, because the spec pins this surface to TerminalAccessCard's plain
useState/optimistic-persist shape. (Revalidation is precisely what SWR would have
given for free, and precisely why most of the hook exists; better spelled out
than implicit.)

Behaviour-preserving: the 20 tests are UNCHANGED and still pass, and the four
mutation checks still kill their mutants against the extracted hook (StrictMode
impure-updater hang, eager resync racing an in-flight PATCH, dirty-draft clobber,
stale-machine write).
